+++ python/ray/rllib/__init__.py
@@ -10,6 +10,7 @@ from ray.tune.registry import register_trainable
 
 from ray.rllib.evaluation.policy_graph import PolicyGraph
 from ray.rllib.evaluation.tf_policy_graph import TFPolicyGraph
+from ray.rllib.evaluation.rollout_worker import RolloutWorker
 from ray.rllib.env.base_env import BaseEnv
 from ray.rllib.env.multi_agent_env import MultiAgentEnv
 from ray.rllib.env.vector_env import VectorEnv
@@ -45,7 +46,7 @@ _register_all()
 __all__ = [
     "PolicyGraph",
     "TFPolicyGraph",
-    "PolicyEvaluator",
+    "RolloutWorker",
     "SampleBatch",
     "BaseEnv",
     "MultiAgentEnv",
