+++ python/ray/rllib/examples/rollout_worker_custom_workflow.py
@@ -1,4 +1,4 @@
-"""Example of using policy evaluator classes directly to implement training.
+"""Example of using rollout worker classes directly to implement training.
 
 Instead of using the built-in Trainer classes provided by RLlib, here we define
 a custom PolicyGraph class and manually coordinate distributed sample
@@ -14,7 +14,8 @@ import gym
 
 import ray
 from ray import tune
-from ray.rllib.evaluation import PolicyGraph, PolicyEvaluator, SampleBatch
+from ray.rllib.policy import Policy
+from ray.rllib.evaluation import RolloutWorker, SampleBatch
 from ray.rllib.evaluation.metrics import collect_metrics
 
 parser = argparse.ArgumentParser()
@@ -66,8 +67,8 @@ def training_workflow(config, reporter):
     env = gym.make("CartPole-v0")
     policy = CustomPolicy(env.observation_space, env.action_space, {})
     workers = [
-        PolicyEvaluator.as_remote().remote(lambda c: gym.make("CartPole-v0"),
-                                           CustomPolicy)
+        RolloutWorker.as_remote().remote(lambda c: gym.make("CartPole-v0"),
+                                         CustomPolicy)
         for _ in range(config["num_workers"])
     ]
