+++ python/ray/rllib/optimizers/sync_batch_replay_optimizer.py
@@ -20,12 +20,11 @@ class SyncBatchReplayOptimizer(PolicyOptimizer):
     This enables RNN support. Does not currently support prioritization."""
 
     def __init__(self,
-                 local_evaluator,
-                 remote_evaluators,
+                 workers,
                  learning_starts=1000,
                  buffer_size=10000,
                  train_batch_size=32):
-        PolicyOptimizer.__init__(self, local_evaluator, remote_evaluators)
+        PolicyOptimizer.__init__(self, workers)
 
         self.replay_starts = learning_starts
         self.max_buffer_size = buffer_size
@@ -45,17 +44,17 @@ class SyncBatchReplayOptimizer(PolicyOptimizer):
     @override(PolicyOptimizer)
     def step(self):
         with self.update_weights_timer:
-            if self.remote_evaluators:
-                weights = ray.put(self.local_evaluator.get_weights())
-                for e in self.remote_evaluators:
+            if self.workers.remote_workers():
+                weights = ray.put(self.workers.local_worker().get_weights())
+                for e in self.workers.remote_workers():
                     e.set_weights.remote(weights)
 
         with self.sample_timer:
-            if self.remote_evaluators:
+            if self.workers.remote_workers():
                 batches = ray_get_and_free(
-                    [e.sample.remote() for e in self.remote_evaluators])
+                    [e.sample.remote() for e in self.workers.remote_workers()])
             else:
-                batches = [self.local_evaluator.sample()]
+                batches = [self.workers.local_worker().sample()]
 
             # Handle everything as if multiagent
             tmp = []
@@ -105,7 +104,7 @@ class SyncBatchReplayOptimizer(PolicyOptimizer):
             samples.append(random.choice(self.replay_buffer))
         samples = SampleBatch.concat_samples(samples)
         with self.grad_timer:
-            info_dict = self.local_evaluator.learn_on_batch(samples)
+            info_dict = self.workers.local_worker().learn_on_batch(samples)
             for policy_id, info in info_dict.items():
                 self.learner_stats[policy_id] = get_learner_stats(info)
             self.grad_timer.push_units_processed(samples.count)
