+++ python/ray/rllib/tests/test_external_multi_agent_env.py
@@ -10,9 +10,9 @@ import unittest
 import ray
 from ray.rllib.agents.pg.pg_policy_graph import PGTFPolicy
 from ray.rllib.optimizers import SyncSamplesOptimizer
-from ray.rllib.evaluation.policy_evaluator import PolicyEvaluator
+from ray.rllib.evaluation.rollout_worker import RolloutWorker
 from ray.rllib.env.external_multi_agent_env import ExternalMultiAgentEnv
-from ray.rllib.tests.test_policy_evaluator import MockPolicyGraph
+from ray.rllib.tests.test_rollout_worker import MockPolicy
 from ray.rllib.tests.test_external_env import make_simple_serving
 from ray.rllib.tests.test_multi_agent_env import BasicMultiAgent, MultiCartpole
 from ray.rllib.evaluation.metrics import collect_metrics
@@ -23,7 +23,7 @@ SimpleMultiServing = make_simple_serving(True, ExternalMultiAgentEnv)
 class TestExternalMultiAgentEnv(unittest.TestCase):
     def testExternalMultiAgentEnvCompleteEpisodes(self):
         agents = 4
-        ev = PolicyEvaluator(
+        ev = RolloutWorker(
             env_creator=lambda _: SimpleMultiServing(BasicMultiAgent(agents)),
             policy_graph=MockPolicyGraph,
             batch_steps=40,
@@ -35,7 +35,7 @@ class TestExternalMultiAgentEnv(unittest.TestCase):
 
     def testExternalMultiAgentEnvTruncateEpisodes(self):
         agents = 4
-        ev = PolicyEvaluator(
+        ev = RolloutWorker(
             env_creator=lambda _: SimpleMultiServing(BasicMultiAgent(agents)),
             policy_graph=MockPolicyGraph,
             batch_steps=40,
@@ -49,7 +49,7 @@ class TestExternalMultiAgentEnv(unittest.TestCase):
         agents = 2
         act_space = gym.spaces.Discrete(2)
         obs_space = gym.spaces.Discrete(2)
-        ev = PolicyEvaluator(
+        ev = RolloutWorker(
             env_creator=lambda _: SimpleMultiServing(BasicMultiAgent(agents)),
             policy_graph={
                 "p0": (MockPolicyGraph, obs_space, act_space, {}),
@@ -70,7 +70,7 @@ class TestExternalMultiAgentEnv(unittest.TestCase):
             policies["pg_{}".format(i)] = (PGTFPolicy, obs_space, act_space,
                                            {})
         policy_ids = list(policies.keys())
-        ev = PolicyEvaluator(
+        ev = RolloutWorker(
             env_creator=lambda _: MultiCartpole(n),
             policy_graph=policies,
             policy_mapping_fn=lambda agent_id: random.choice(policy_ids),
