+++ python/ray/rllib/tests/test_multi_agent_env.py
@@ -12,11 +12,11 @@ from ray.rllib.agents.pg.pg_policy_graph import PGTFPolicy
 from ray.rllib.agents.dqn.dqn_policy_graph import DQNPolicyGraph
 from ray.rllib.optimizers import (SyncSamplesOptimizer, SyncReplayOptimizer,
                                   AsyncGradientsOptimizer)
-from ray.rllib.tests.test_policy_evaluator import (MockEnv, MockEnv2,
-                                                   MockPolicyGraph)
-from ray.rllib.evaluation.policy_evaluator import PolicyEvaluator
-from ray.rllib.evaluation.policy_graph import PolicyGraph
+from ray.rllib.tests.test_rollout_worker import (MockEnv, MockEnv2, MockPolicy)
+from ray.rllib.evaluation.rollout_worker import RolloutWorker
+from ray.rllib.policy.policy import Policy
 from ray.rllib.evaluation.metrics import collect_metrics
+from ray.rllib.evaluation.worker_set import WorkerSet
 from ray.rllib.env.base_env import _MultiAgentEnvToBaseEnv
 from ray.rllib.env.multi_agent_env import MultiAgentEnv
 from ray.tune.registry import register_env
@@ -327,7 +327,7 @@ class TestMultiAgentEnv(unittest.TestCase):
     def testMultiAgentSample(self):
         act_space = gym.spaces.Discrete(2)
         obs_space = gym.spaces.Discrete(2)
-        ev = PolicyEvaluator(
+        ev = RolloutWorker(
             env_creator=lambda _: BasicMultiAgent(5),
             policy_graph={
                 "p0": (MockPolicyGraph, obs_space, act_space, {}),
@@ -345,7 +345,7 @@ class TestMultiAgentEnv(unittest.TestCase):
     def testMultiAgentSampleSyncRemote(self):
         act_space = gym.spaces.Discrete(2)
         obs_space = gym.spaces.Discrete(2)
-        ev = PolicyEvaluator(
+        ev = RolloutWorker(
             env_creator=lambda _: BasicMultiAgent(5),
             policy_graph={
                 "p0": (MockPolicyGraph, obs_space, act_space, {}),
@@ -362,7 +362,7 @@ class TestMultiAgentEnv(unittest.TestCase):
     def testMultiAgentSampleAsyncRemote(self):
         act_space = gym.spaces.Discrete(2)
         obs_space = gym.spaces.Discrete(2)
-        ev = PolicyEvaluator(
+        ev = RolloutWorker(
             env_creator=lambda _: BasicMultiAgent(5),
             policy_graph={
                 "p0": (MockPolicyGraph, obs_space, act_space, {}),
@@ -378,7 +378,7 @@ class TestMultiAgentEnv(unittest.TestCase):
     def testMultiAgentSampleWithHorizon(self):
         act_space = gym.spaces.Discrete(2)
         obs_space = gym.spaces.Discrete(2)
-        ev = PolicyEvaluator(
+        ev = RolloutWorker(
             env_creator=lambda _: BasicMultiAgent(5),
             policy_graph={
                 "p0": (MockPolicyGraph, obs_space, act_space, {}),
@@ -393,7 +393,7 @@ class TestMultiAgentEnv(unittest.TestCase):
     def testSampleFromEarlyDoneEnv(self):
         act_space = gym.spaces.Discrete(2)
         obs_space = gym.spaces.Discrete(2)
-        ev = PolicyEvaluator(
+        ev = RolloutWorker(
             env_creator=lambda _: EarlyDoneMultiAgent(),
             policy_graph={
                 "p0": (MockPolicyGraph, obs_space, act_space, {}),
@@ -409,7 +409,7 @@ class TestMultiAgentEnv(unittest.TestCase):
     def testMultiAgentSampleRoundRobin(self):
         act_space = gym.spaces.Discrete(2)
         obs_space = gym.spaces.Discrete(10)
-        ev = PolicyEvaluator(
+        ev = RolloutWorker(
             env_creator=lambda _: RoundRobinMultiAgent(5, increment_obs=True),
             policy_graph={
                 "p0": (MockPolicyGraph, obs_space, act_space, {}),
@@ -458,7 +458,7 @@ class TestMultiAgentEnv(unittest.TestCase):
             def get_initial_state(self):
                 return [{}]  # empty dict
 
-        ev = PolicyEvaluator(
+        ev = RolloutWorker(
             env_creator=lambda _: gym.make("CartPole-v0"),
             policy_graph=StatefulPolicyGraph,
             batch_steps=5)
@@ -503,7 +503,7 @@ class TestMultiAgentEnv(unittest.TestCase):
         single_env = gym.make("CartPole-v0")
         obs_space = single_env.observation_space
         act_space = single_env.action_space
-        ev = PolicyEvaluator(
+        ev = RolloutWorker(
             env_creator=lambda _: MultiCartpole(2),
             policy_graph={
                 "p0": (ModelBasedPolicyGraph, obs_space, act_space, {}),
@@ -587,7 +587,7 @@ class TestMultiAgentEnv(unittest.TestCase):
                 "p1": (PGTFPolicy, obs_space, act_space, {}),
                 "p2": (DQNPolicyGraph, obs_space, act_space, dqn_config),
             }
-        ev = PolicyEvaluator(
+        worker = RolloutWorker(
             env_creator=lambda _: MultiCartpole(n),
             policy_graph=policies,
             policy_mapping_fn=lambda agent_id: ["p1", "p2"][agent_id % 2],
@@ -597,25 +597,30 @@ class TestMultiAgentEnv(unittest.TestCase):
             def policy_mapper(agent_id):
                 return ["p1", "p2"][agent_id % 2]
 
-            remote_evs = [
-                PolicyEvaluator.as_remote().remote(
+            remote_workers = [
+                RolloutWorker.as_remote().remote(
                     env_creator=lambda _: MultiCartpole(n),
                     policy_graph=policies,
                     policy_mapping_fn=policy_mapper,
                     batch_steps=50)
             ]
         else:
-            remote_evs = []
-        optimizer = optimizer_cls(ev, remote_evs)
+            remote_workers = []
+        workers = WorkerSet._from_existing(worker, remote_workers)
+        optimizer = optimizer_cls(workers)
         for i in range(200):
-            ev.foreach_policy(lambda p, _: p.set_epsilon(
+            worker.foreach_policy(lambda p, _: p.set_epsilon(
                 max(0.02, 1 - i * .02))
                               if isinstance(p, DQNPolicyGraph) else None)
             optimizer.step()
-            result = collect_metrics(ev, remote_evs)
+            result = collect_metrics(worker, remote_workers)
             if i % 20 == 0:
-                ev.foreach_policy(lambda p, _: p.update_target() if isinstance(
-                    p, DQNPolicyGraph) else None)
+
+                def do_update(p):
+                    if isinstance(p, DQNTFPolicy):
+                        p.update_target()
+
+                worker.foreach_policy(lambda p, _: do_update(p))
                 print("Iter {}, rew {}".format(i,
                                                result["policy_reward_mean"]))
                 print("Total reward", result["episode_reward_mean"])
@@ -643,15 +648,16 @@ class TestMultiAgentEnv(unittest.TestCase):
             policies["pg_{}".format(i)] = (PGTFPolicy, obs_space, act_space,
                                            {})
         policy_ids = list(policies.keys())
-        ev = PolicyEvaluator(
+        worker = RolloutWorker(
             env_creator=lambda _: MultiCartpole(n),
             policy_graph=policies,
             policy_mapping_fn=lambda agent_id: random.choice(policy_ids),
             batch_steps=100)
-        optimizer = SyncSamplesOptimizer(ev, [])
+        workers = WorkerSet._from_existing(worker, [])
+        optimizer = SyncSamplesOptimizer(workers)
         for i in range(100):
             optimizer.step()
-            result = collect_metrics(ev)
+            result = collect_metrics(worker)
             print("Iteration {}, rew {}".format(i,
                                                 result["policy_reward_mean"]))
             print("Total reward", result["episode_reward_mean"])
